tests/requirements.txt
# Perfect21测试依赖包

# 测试框架
pytest>=7.0.0
pytest-asyncio>=0.21.0
pytest-cov>=4.0.0
pytest-html>=3.1.0
pytest-xdist>=3.0.0
pytest-mock>=3.10.0
pytest-timeout>=2.1.0
pytest-benchmark>=4.0.0

# FastAPI测试
httpx>=0.24.0
fastapi[all]>=0.100.0
starlette>=0.27.0

# 数据库测试
pytest-postgresql>=4.1.0
pytest-redis>=3.0.0
sqlalchemy>=2.0.0
sqlite3  # Python内置

# 性能测试
aiohttp>=3.8.0
psutil>=5.9.0
locust>=2.15.0

# 安全测试
cryptography>=41.0.0
bcrypt>=4.0.0
pyjwt>=2.8.0

# Mock 和 Fixture
factory-boy>=3.2.0
faker>=19.0.0
responses>=0.23.0

# 报告和分析
allure-pytest>=2.13.0
pytest-json-report>=1.5.0
pytest-metadata>=3.0.0

# 代码质量
flake8>=6.0.0
black>=23.0.0
isort>=5.12.0
mypy>=1.0.0

# 日志和调试
structlog>=23.0.0
coloredlogs>=15.0.0

# 其他工具
pyyaml>=6.0
toml>=0.10.2
pathlib>=1.0.1
